requirements.txt
Cython==0.29.34
aiohttp>=3.0.0,<3.6.0;python_version<"3.8"
aiohttp>=3.6.0,<4.0.0;python_version>="3.8"
asyncpg==0.27.0
cffi==1.15.1
chardet==3.0.4
jsonschema==2.6.0
multidict==4.5.2
pycparser==2.18
pycryptodome==3.6.6
PyJWT==2.6.0
python-dateutil==2.8.2
PyYaml==5.1
six==1.11.0
ujson==1.35
yarl==1.2.6
zope.interface==5.1.0
argon2-cffi==21.3.0
backoff==1.10.0
prometheus-client==0.8.0
brotli==1.0.9
typing_extensions==3.7.4.3
types-PyYAML==6.0.12.9
types-docutils==0.19.1.7
types-python-dateutil==2.8.19.12
types-setuptools==67.6.0.7
types-ujson==5.7.0.1
